requirements.txt
joblib==1.3.2
libsvm==3.23.0.4
matplotlib==3.8.2
mysql-connector-python==8.4.0
nltk==3.8.1
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0
streamlit==1.34.0
streamlit-option-menu==0.3.12
wordcloud==1.9.3